fixup! Add decorators for /receive and /receive_out_of_band views
